Update to v3
