+++ package.json
@@ -1,22 +1,26 @@
 {
   "name": "@okteto-website/docs",
-  "version": "2.0.0-alpha2",
+  "version": "3.1.0",
   "private": true,
   "dependencies": {
-    "@docusaurus/core": "^2.0.0-beta.17",
-    "@docusaurus/plugin-client-redirects": "^2.0.0-beta.17",
-    "@docusaurus/plugin-content-pages": "^2.3.1",
-    "@docusaurus/preset-classic": "^2.0.0-beta.17",
+    "@docusaurus/core": "^3.1.0",
+    "@docusaurus/plugin-client-redirects": "^3.1.0",
+    "@docusaurus/plugin-content-pages": "^3.1.0",
+    "@docusaurus/preset-classic": "^3.1.0",
+    "@mdx-js/react": "^3.0.0",
     "@sanity/client": "^3.2.0",
     "classnames": "^2.3.1",
     "docusaurus-gtm-plugin": "^0.0.2",
     "docusaurus-plugin-copy": "^0.1.1",
-    "docusaurus-plugin-sass": "^0.1.12",
+    "docusaurus-plugin-sass": "^0.2.5",
     "lodash.trimend": "^4.5.1",
     "mdx-observable": "^0.2.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "sass": "^1.35.2"
+  },
+  "engines" : { 
+    "node" : ">=18.0"
   },
   "browserslist": {
     "production": [
