--- src/theme/CodeBlock/styles.module.css
@@ -1,76 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-.codeBlockContainer {
-  margin-bottom: var(--ifm-leading);
-  border-radius: var(--ifm-global-radius);
-  box-shadow: var(--ifm-global-shadow-lw);
-}
-
-.codeBlockContent {
-  position: relative;
-  /* rtl:ignore */
-  direction: ltr;
-}
-
-.codeBlockTitle {
-  border-bottom: 1px solid var(--ifm-color-emphasis-300);
-  font-size: var(--ifm-code-font-size);
-  font-weight: 500;
-  padding: 0.75rem var(--ifm-pre-padding);
-  border-top-left-radius: var(--ifm-global-radius);
-  border-top-right-radius: var(--ifm-global-radius);
-}
-
-.codeBlock {
-  margin: 0;
-  padding: 0;
-  border-radius: var(--ifm-global-radius);
-}
-
-.codeBlockTitle + .codeBlockContent .codeBlock {
-  border-top-left-radius: 0;
-  border-top-right-radius: 0;
-}
-
-.codeBlockStandalone {
-  padding: 0;
-  border-radius: var(--ifm-global-radius);
-}
-
-.copyButton {
-  background: rgb(0 0 0 / 30%);
-  border-radius: var(--ifm-global-radius);
-  color: var(--ifm-color-white);
-  opacity: 0;
-  user-select: none;
-  padding: 0.4rem 0.5rem;
-  position: absolute;
-  right: calc(var(--ifm-pre-padding) / 2);
-  top: calc(var(--ifm-pre-padding) / 2);
-  transition: opacity 200ms ease-in-out;
-}
-
-.copyButton:focus,
-.codeBlockContent:hover > .copyButton,
-.codeBlockTitle:hover + .codeBlockContent .copyButton {
-  opacity: 1;
-}
-
-.codeBlockLines {
-  font: inherit;
-  /* rtl:ignore */
-  float: left;
-  min-width: 100%;
-  padding: var(--ifm-pre-padding);
-}
-
-@media print {
-  .codeBlockLines {
-    white-space: pre-wrap;
-  }
-}
