+++ versioned_docs/version-1.16/core/using-okteto-cli.mdx
@@ -17,7 +17,6 @@ Learn more about the Okteto CLI
 - [FAQs](reference/faqs.mdx)
 - [Known Issues](reference/known-issues.mdx)
 
-
 ## Deployment Manifests
 
 You need one of the following deployment manifests for your application so that Okteto CLI knows how to launch a Dev Environment for it. Okteto CLI will analyze your existing manifests (Helm, Kubernetes, etc.) and help you write an [Okteto Manifest](#okteto-manifest) for your application. It will then use this Okteto Manifest instead of your existing deployment manifests to spin up Dev Environments.
@@ -33,7 +32,6 @@ Okteto CLI will look for deployment manifests for your application in the follow
 1. [Kubernetes manifests](https://kubernetes.io/docs/concepts/cluster-administration/manage-deployment/) file: if there is a `manifests.yaml`, `kubernetes.yaml`, or `k8s.yaml` file, Okteto will detect it and run `kubectl apply` to deploy your development environment
 
 You can also explicitly define the location of the Okteto Manifest using the `-f` flag on any Okteto CLI command.
-
 
 ## Okteto manifest
 
@@ -67,6 +65,7 @@ If you want to learn more about what all you can configure in this manifest, hea
 #### Default Environment Variables
 
 The following environment variables are automatically populated to be referred in your Okteto Manifest or deploy/destroy commands:
+
 - `OKTETO_DOMAIN`: the domain where Okteto exposes your application endpoints.
 - `OKTETO_NAMESPACE`: the namespace where your application is going to be installed.
 - `OKTETO_USERNAME`: your username in Okteto.
@@ -77,6 +76,7 @@ The following environment variables are automatically populated to be referred i
 #### Built-in Environment Variables for Images stored in the Okteto Registry
 
 You can refer to the images defined in the [build](reference/okteto-manifest.mdx/#build-object-optional) section of the Okteto Manifest using the following environment variables:
+
 - `OKTETO_BUILD_<image>_IMAGE`: the full image reference
 - `OKTETO_BUILD_<image>_REGISTRY`: the registry URL where the image was pushed
 - `OKTETO_BUILD_<image>_REPOSITORY`: the name of the image that was pushed
@@ -97,11 +97,12 @@ Okteto will build a random image tag like `registry.okteto.example.com/cindy/hel
 - `OKTETO_BUILD_HELLO_WORLD_REPOSITORY`: cindy/hello-world
 - `OKTETO_BUILD_HELLO_WORLD_SHA`: f56119a37b@sha256:xxx
 
-> if an `<image>` name includes the symbol `-`, in the corresponding `OKTETO_BUILD_<image>` environment variables, the symbol `-` will be replaced by `_`.  For example, if `<image>` is `name-with-hyphen`, the environment variable names will start with `$OKTETO_BUILD_NAME_WITH_HYPHEN_`.
+> if an `<image>` name includes the symbol `-`, in the corresponding `OKTETO_BUILD_<image>` environment variables, the symbol `-` will be replaced by `_`. For example, if `<image>` is `name-with-hyphen`, the environment variable names will start with `$OKTETO_BUILD_NAME_WITH_HYPHEN_`.
 
 #### Built-in Environment Variables for Dependencies
 
 You can refer to the variables defined by your [dependencies](reference/okteto-manifest.mdx/#dependencies-object-optional) using the following environment variables:
+
 - `OKTETO_DEPENDENCY_<dependency>_VARIABLE_<variable>`: the value of a variable of a dependency.
 
 For example, if you have the following dependency in your Okteto Manifest:
